* commit 'eaf7e5611aa07f6d78f060208369048a1ae099d4':
  update manual tests
  service discovery for broker
  update tests
  output spdlogs as json
  fix
  http health check for nginx
  correct windows tests
  daemon off for nginx
  disable nginx logs
  access discovery service via nginx
  start nginx

+++ tests/manual/performance_producer_receiver/test.sh
@@ -4,6 +4,14 @@ set -e
 
 trap Cleanup EXIT
 
+Cleanup() {
+set +e
+ssh ${service_node} rm -f ${service_dir}/files/*
+ssh ${service_node} killall receiver
+ssh ${service_node} killall asapo-discovery
+ssh ${service_node} docker rm -f -v mongo
+}
+
 
 # starts receiver on $service_node
 # runs producer with various file sizes from $worker_node and measures performance
@@ -48,20 +56,24 @@ cat receiver.json |
           end
          ) |
       from_entries" > receiver_tmp.json
+
 cat discovery.json |
   jq "to_entries |
        map(if .key == \"Port\"
           then . + {value:${discovery_port}}
-          elif .key == \"Endpoints\"
-          then . + {value:[\"${service_node}:${receiver_port}\"]}
           else .
           end
          ) |
       from_entries" > discovery_tmp.json
 
-scp discovery_tmp.json ${service_node}:${service_dir}/discovery.json
+cat discovery.json | jq ".Port = ${discovery_port}" > discovery_tmp.json
+
+cat discovery_tmp.json | jq ".Receiver.StaticEndpoints = [\"${service_node}:${receiver_port}\"]" > discovery_tmp1.json
+
+
+scp discovery_tmp1.json ${service_node}:${service_dir}/discovery.json
 scp receiver_tmp.json ${service_node}:${service_dir}/receiver.json
-rm discovery_tmp.json receiver_tmp.json
+rm discovery_tmp*.json receiver_tmp.json
 ssh ${service_node} "bash -c 'cd ${service_dir}; nohup ./receiver receiver.json &> ${service_dir}/receiver.log &'"
 ssh ${service_node} "bash -c 'cd ${service_dir}; nohup ./asapo-discovery -config discovery.json &> ${service_dir}/discovery.log &'"
 
@@ -70,8 +82,11 @@ for size  in 100 1000 10000
 do
 ssh ${service_node} docker run -d -p 27017:27017 --name mongo mongo
 echo ===================================================================
-ssh ${worker_node} ${worker_dir}/dummy-data-producer ${service_ip}:${discovery_port} ${size} 1000 8 0
-ssh ${service_node} rm -f ${service_dir}/files/*
+ssh ${worker_node} ${worker_dir}/dummy-data-producer ${service_ip}:8400 ${size} 1000 8 0
+if [ "$1" == "true" ]
+then
+    ssh ${service_node} rm -f ${service_dir}/files/*
+fi
 ssh ${service_node} docker rm -f -v mongo
 done
 ssh ${service_node} killall receiver
